+++ test/configuration/index.spec.ts
@@ -5,18 +5,18 @@ import Ajv from 'ajv';
 
 import ConfigurationSingleton, {
 	Configuration,
-	defaultOptions
+	defaultOptions,
 } from '../../src/configuration';
 import { IConfigOptions } from '../../src/interfaces/IConfigOptions';
 
 describe('test configuration', () => {
 	const options: IConfigOptions = {
-		configDir: 'test/data'
+		configDir: 'test/data',
 	};
 
 	it('test configuration default initialization', () => {
 		const config = new Configuration({
-			configDir: 'test/data'
+			configDir: 'test/data',
 		});
 		expect(config).to.be.not.null;
 		expect(config).to.be.not.undefined;
@@ -32,17 +32,18 @@ describe('test configuration', () => {
 	});
 
 	it('test assignmment of default values', () => {
+		const defaultSample = 'defaultSample';
 		const options: IConfigOptions = {
 			configDir: 'test/data',
 			notFoundValue: false,
-			throwOnError: false
+			throwOnError: false,
 		};
 		const config = new Configuration(options);
 		// this will be set as default from schema definition
 		expect(
 			config.get('DefaultSample'),
 			'default value has been applied'
-		).to.be.equal('defaultSample');
+		).to.be.equal(defaultSample);
 		// this will be removed because not in schema but in ENV
 		expect(options.notFoundValue).to.be.equal(false); // default is null use something different here
 		expect(
@@ -69,6 +70,23 @@ describe('test configuration', () => {
 			true
 		);
 		expect(config.get('Boolean'), 'get Boolean').to.be.equal(false);
+		expect(config.get('DefaultBoolean'), 'get default Boolean').to.be.equal(
+			true
+		);
+		expect(
+			config.remove('DefaultBoolean'),
+			'remove default Boolean does nothing'
+		).to.be.equal(true);
+		expect(config.get('DefaultBoolean'), 'get default Boolean').to.be.equal(
+			true
+		);
+		expect(
+			config.set('DefaultBoolean', false),
+			'set default Boolean'
+		).to.be.equal(true);
+		expect(config.get('DefaultBoolean'), 'get default Boolean').to.be.equal(
+			false
+		);
 		expect(config.getErrors(), 'no errors exist').to.be.null;
 		expect(config.set('Boolean', true), 'boolean re-assignment').to.be.equal(
 			true
@@ -82,6 +100,20 @@ describe('test configuration', () => {
 			true
 		);
 		expect(config.get('String'), 'get String').to.be.equal('bar');
+		// removal of defaults should keep defaults
+		const defaultSample = 'defaultSample';
+		const defaultSampleEmtyString = '';
+		expect(config.get('DefaultSample')).to.be.equal(defaultSample);
+		expect(config.remove('DefaultSample')).to.be.equal(true);
+		expect(config.get('DefaultSample')).to.be.equal(defaultSample);
+		// like assigning null does not change anything
+		expect(config.set('DefaultSample', null)).to.be.equal(true);
+		expect(config.get('DefaultSample')).to.be.equal(defaultSampleEmtyString);
+		// but empty string is allowed
+		expect(config.set('DefaultSample', defaultSampleEmtyString)).to.be.equal(
+			true
+		);
+		expect(config.get('DefaultSample')).to.be.equal(defaultSampleEmtyString);
 		expect(config.getErrors(), 'no errors exist').to.be.null;
 	});
 
@@ -150,7 +182,7 @@ describe('test configuration', () => {
 			schemaFileName: 'dot.schema.json',
 			configDir: 'test/data',
 			envDir: 'test/data',
-			debug: true
+			debug: true,
 		};
 		const { dotNotationSeparator } = defaultOptions;
 		const configNestedFoo = 'Nested' + dotNotationSeparator + 'foo';
@@ -161,9 +193,9 @@ describe('test configuration', () => {
 				Sample: 'sample',
 				Nested: {
 					foo: 'foo',
-					bar: 'bar'
+					bar: 'bar',
 				},
-				Very: { Nested: { Value: 'value' } }
+				Very: { Nested: { Value: 'value' } },
 			};
 
 			const dotted = dot.dot(sample);
@@ -178,7 +210,7 @@ describe('test configuration', () => {
 				Sample: 'sample',
 				'Nested.bar': 'bar',
 				'Nested.foo': 'foo',
-				'Very.Nested.Value': 'value'
+				'Very.Nested.Value': 'value',
 			};
 			const objected: any = dot.object(sample);
 			expect(objected.Sample).to.be.equal('sample');
@@ -258,6 +290,49 @@ describe('test configuration', () => {
 		});
 	});
 
+	describe('modifying function access is restricted in production', () => {
+		it('updating configuration works in test mode (not production mode)', () => {
+			const { NODE_ENV: NODE_ENV_BEFORE } = process.env;
+			expect(NODE_ENV_BEFORE, 'must set test').to.be.equal('test');
+			const config = new Configuration(options);
+			// test all public methods are not failing
+			expect(config.has('DefaultSample')).to.be.true;
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+			const currentConfig = config.toObject();
+			expect(currentConfig).to.haveOwnProperty('DefaultSample');
+			config.set('DefaultSample', 'foo');
+			expect(config.get('DefaultSample')).to.be.equal('foo');
+			config.update({ DefaultSample: 'bar' });
+			expect(config.get('DefaultSample')).to.be.equal('bar');
+			config.reset(currentConfig);
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+			config.remove('DefaultSample'); // does not change 'DefaultSample', but is executed
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+		});
+		it('updating configuration works not in production mode', () => {
+			const { NODE_ENV: NODE_ENV_BEFORE } = process.env;
+			process.env.NODE_ENV = 'production';
+			const { NODE_ENV } = process.env;
+			expect(NODE_ENV, 'must set production').to.be.equal('production');
+			const config = new Configuration(options);
+			expect(config.has('DefaultSample')).to.be.true;
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+			const currentConfig = config.toObject();
+			expect(currentConfig).to.haveOwnProperty('DefaultSample');
+			// test all other public methods are failing
+			const regExp = new RegExp('changes during runtime are not allowed');
+			expect(() => config.set('DefaultSample', 'foo')).throws(regExp);
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+			expect(() => config.update({ DefaultSample: 'bar' })).throws(regExp);
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+			expect(() => config.reset(currentConfig)).throws(regExp);
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+			expect(() => config.remove('DefaultSample')).throws(regExp); // does not change 'DefaultSample', but is executed
+			expect(config.get('DefaultSample')).to.be.equal('defaultSample');
+			process.env.NODE_ENV = NODE_ENV_BEFORE;
+		});
+	});
+
 	describe('schema dependencies', () => {
 		it('property-value based dependency (feature-flag condition)', () => {
 			// Specifying dependencies this way does not work when the ajv option {removeAdditional: 'all'} is set,
@@ -274,10 +349,10 @@ describe('test configuration', () => {
 
 			const url = 'http://example.tld';
 			expect(() => config.set('FEATURE_FLAG', true)).to.throw;
-			expect(config.update({ 'FEATURE_OPTION': url, 'FEATURE_FLAG': true })).to.be.true;
+			expect(config.update({ FEATURE_OPTION: url, FEATURE_FLAG: true })).to.be
+				.true;
 			expect(config.get('FEATURE_FLAG')).to.be.true;
 			expect(config.get('FEATURE_OPTION')).to.be.equal(url);
-
 		});
 	});
 
@@ -285,10 +360,10 @@ describe('test configuration', () => {
 		it('should reset given values', () => {
 			const config = new Configuration({
 				schemaFileName: 'default.schema.json',
-				configDir: 'test/data'
+				configDir: 'test/data',
 			});
 			const before = config.toObject();
-			expect(Object.keys(before).length).to.be.equal(7);
+			expect(Object.keys(before).length).to.be.equal(9);
 			expect(before['Domain']).to.be.equal('localhost');
 			config.set('Domain', 'otherdomain.tld');
 			expect(config.get('Domain')).to.be.equal('otherdomain.tld');
@@ -298,10 +373,10 @@ describe('test configuration', () => {
 		it('should remove values not set before', () => {
 			const config = new Configuration({
 				schemaFileName: 'default.schema.json',
-				configDir: 'test/data'
+				configDir: 'test/data',
 			});
 			const before = config.toObject();
-			expect(Object.keys(before).length).to.be.equal(7);
+			expect(Object.keys(before).length).to.be.equal(9);
 			expect('String' in before).to.be.false;
 			expect(() => config.get('String')).to.throw;
 			config.set('String', 'newValueNotDefinedBefore');
@@ -315,7 +390,7 @@ describe('test configuration', () => {
 		it('remove single manually added key', () => {
 			const config = new Configuration({
 				schemaFileName: 'default.schema.json',
-				configDir: 'test/data'
+				configDir: 'test/data',
 			});
 			config.set('String', 'sample');
 			const before = config.toObject();
@@ -330,7 +405,7 @@ describe('test configuration', () => {
 		it('remove multiple keys', () => {
 			const config = new Configuration({
 				schemaFileName: 'default.schema.json',
-				configDir: 'test/data'
+				configDir: 'test/data',
 			});
 			config.set('String', 'sample');
 			config.set('Boolean', 'true');
@@ -349,7 +424,7 @@ describe('test configuration', () => {
 		it('remove required key fails', () => {
 			const config = new Configuration({
 				schemaFileName: 'default.schema.json',
-				configDir: 'test/data'
+				configDir: 'test/data',
 			});
 			config.set('Domain', 'sample.de');
 			const before = config.toObject();
@@ -364,42 +439,40 @@ describe('test configuration', () => {
 		const ajv = new Ajv({
 			removeAdditional: true,
 			useDefaults: true,
-			coerceTypes: 'array'
+			coerceTypes: 'array',
 		});
 
 		describe('if/then', () => {
 			const schema = {
-				'title': 'dependency test schema',
-				'type': 'object',
-				'description': 'this schema declares different dependency methodologies based on specific values defined',
-				'additionalProperties': false,
-				'properties': {
-					'FEATURE_FLAG': {
-						'type': 'boolean',
-						'default': false
+				title: 'dependency test schema',
+				type: 'object',
+				description:
+					'this schema declares different dependency methodologies based on specific values defined',
+				additionalProperties: false,
+				properties: {
+					FEATURE_FLAG: {
+						type: 'boolean',
+						default: false,
 					},
-					'FEATURE_OPTION': {
-						'type': 'string',
-						'format': 'uri'
+					FEATURE_OPTION: {
+						type: 'string',
+						format: 'uri',
 					},
-					'OTHER_FEATURE_OPTION': {
-						'type': 'number',
-						'default': 42
-					}
+					OTHER_FEATURE_OPTION: {
+						type: 'number',
+						default: 42,
+					},
 				},
-				'if': {
-					'properties': {
-						'FEATURE_FLAG': {
-							'const': true
-						}
-					}
+				if: {
+					properties: {
+						FEATURE_FLAG: {
+							const: true,
+						},
+					},
 				},
-				'then': {
-					'required': [
-						'FEATURE_OPTION',
-						'OTHER_FEATURE_OPTION'
-					]
-				}
+				then: {
+					required: ['FEATURE_OPTION', 'OTHER_FEATURE_OPTION'],
+				},
 			};
 			const validate = ajv.compile(schema);
 
@@ -410,33 +483,33 @@ describe('test configuration', () => {
 
 			it('valid with all options', () => {
 				const valid = validate({
-					'FEATURE_FLAG': true,
-					'FEATURE_OPTION': 'http://asd.de',
-					'OTHER_FEATURE_OPTION': 12
+					FEATURE_FLAG: true,
+					FEATURE_OPTION: 'http://asd.de',
+					OTHER_FEATURE_OPTION: 12,
 				});
 				expect(valid).to.be.true;
 			});
 
 			it('invalid without option', () => {
 				const valid = validate({
-					'FEATURE_FLAG': true,
-					'OTHER_FEATURE_OPTION': 12
+					FEATURE_FLAG: true,
+					OTHER_FEATURE_OPTION: 12,
 				});
 				expect(valid).to.be.false;
 			});
 
 			it('valid if false', () => {
 				const valid = validate({
-					'FEATURE_FLAG': false,
-					'OTHER_FEATURE_OPTION': 12
+					FEATURE_FLAG: false,
+					OTHER_FEATURE_OPTION: 12,
 				});
 				expect(valid).to.be.true;
 			});
 
 			it('valid with defaults', () => {
 				const valid = validate({
-					'FEATURE_FLAG': true,
-					'FEATURE_OPTION': 'http://asd.de',
+					FEATURE_FLAG: true,
+					FEATURE_OPTION: 'http://asd.de',
 				});
 				expect(valid).to.be.true;
 			});
@@ -444,46 +517,45 @@ describe('test configuration', () => {
 
 		describe('if/then $ref', () => {
 			const schema = {
-				'title': 'dependency test schema',
-				'type': 'object',
-				'description': 'this schema declares different dependency methodologies based on specific values defined',
-				'additionalProperties': false,
-				'properties': {
-					'FEATURE_FLAG': {
-						'type': 'boolean',
-						'default': false
+				title: 'dependency test schema',
+				type: 'object',
+				description:
+					'this schema declares different dependency methodologies based on specific values defined',
+				additionalProperties: false,
+				properties: {
+					FEATURE_FLAG: {
+						type: 'boolean',
+						default: false,
 					},
-					'FEATURE_OPTION': {
-						'type': 'string',
-						'format': 'uri'
+					FEATURE_OPTION: {
+						type: 'string',
+						format: 'uri',
 					},
-					'OTHER_FEATURE_OPTION': {
-						'type': 'number',
-						'default': 42
-					}
+					OTHER_FEATURE_OPTION: {
+						type: 'number',
+						default: 42,
+					},
 				},
-				'allOf': [
+				allOf: [
 					{
-						'$ref': '#/definitions/require_feature'
-					}
+						$ref: '#/definitions/require_feature',
+					},
 				],
-				'definitions': {
-					'require_feature': {
-						'if': {
-							'properties': {
-								'FEATURE_FLAG': {
-									'const': true
-								}
-							}
+				definitions: {
+					// eslint-disable-next-line @typescript-eslint/camelcase
+					require_feature: {
+						if: {
+							properties: {
+								FEATURE_FLAG: {
+									const: true,
+								},
+							},
 						},
-						'then': {
-							'required': [
-								'FEATURE_OPTION',
-								'OTHER_FEATURE_OPTION'
-							]
-						}
-					}
-				}
+						then: {
+							required: ['FEATURE_OPTION', 'OTHER_FEATURE_OPTION'],
+						},
+					},
+				},
 			};
 			const validate = ajv.compile(schema);
 
@@ -494,33 +566,33 @@ describe('test configuration', () => {
 
 			it('valid with all options', () => {
 				const valid = validate({
-					'FEATURE_FLAG': true,
-					'FEATURE_OPTION': 'http://asd.de',
-					'OTHER_FEATURE_OPTION': 12
+					FEATURE_FLAG: true,
+					FEATURE_OPTION: 'http://asd.de',
+					OTHER_FEATURE_OPTION: 12,
 				});
 				expect(valid).to.be.true;
 			});
 
 			it('invalid without option', () => {
 				const valid = validate({
-					'FEATURE_FLAG': true,
-					'OTHER_FEATURE_OPTION': 12
+					FEATURE_FLAG: true,
+					OTHER_FEATURE_OPTION: 12,
 				});
 				expect(valid).to.be.false;
 			});
 
 			it('valid if false', () => {
 				const valid = validate({
-					'FEATURE_FLAG': false,
-					'OTHER_FEATURE_OPTION': 12
+					FEATURE_FLAG: false,
+					OTHER_FEATURE_OPTION: 12,
 				});
 				expect(valid).to.be.true;
 			});
 
 			it('valid with defaults', () => {
 				const valid = validate({
-					'FEATURE_FLAG': true,
-					'FEATURE_OPTION': 'http://asd.de',
+					FEATURE_FLAG: true,
+					FEATURE_OPTION: 'http://asd.de',
 				});
 				expect(valid).to.be.true;
 			});
@@ -530,63 +602,62 @@ describe('test configuration', () => {
 			// good to know: conditional validation is always true when the properties is `undefined`
 			// https://github.com/epoberezkin/ajv/issues/913
 			const schema = {
-				'title': 'dependency test schema',
-				'type': 'object',
-				'description': 'this schema declares different dependency methodologies based on specific values defined',
-				'additionalProperties': false,
-				'properties': {
-					'SERVICE': {
-						'type': 'string',
-						'enum': ['none', 'SERVICE_A', 'SERVICE_B'],
-						'default': 'none'
+				title: 'dependency test schema',
+				type: 'object',
+				description:
+					'this schema declares different dependency methodologies based on specific values defined',
+				additionalProperties: false,
+				properties: {
+					SERVICE: {
+						type: 'string',
+						enum: ['none', 'SERVICE_A', 'SERVICE_B'],
+						default: 'none',
 					},
-					'SERVICE_A_URL': {
-						'type': 'string',
-						'format': 'uri'
+					SERVICE_A_URL: {
+						type: 'string',
+						format: 'uri',
 					},
-					'SERVICE_B_URL': {
-						'type': 'string',
-						'format': 'uri'
-					}
+					SERVICE_B_URL: {
+						type: 'string',
+						format: 'uri',
+					},
 				},
-				'allOf': [
+				allOf: [
 					{
-						'$ref': '#/definitions/require_service_a_url',
+						$ref: '#/definitions/require_service_a_url',
 					},
 					{
-						'$ref': '#/definitions/require_service_b_url',
+						$ref: '#/definitions/require_service_b_url',
 					},
 				],
-				'definitions': {
-					'require_service_a_url': {
-						'if': {
-							'properties': {
-								'SERVICE': {
-									'const': 'SERVICE_A'
-								}
-							}
+				definitions: {
+					// eslint-disable-next-line @typescript-eslint/camelcase
+					require_service_a_url: {
+						if: {
+							properties: {
+								SERVICE: {
+									const: 'SERVICE_A',
+								},
+							},
 						},
-						'then': {
-							'required': [
-								'SERVICE_A_URL',
-							]
-						}
+						then: {
+							required: ['SERVICE_A_URL'],
+						},
 					},
-					'require_service_b_url': {
-						'if': {
-							'properties': {
-								'SERVICE': {
-									'const': 'SERVICE_B'
-								}
-							}
+					// eslint-disable-next-line @typescript-eslint/camelcase
+					require_service_b_url: {
+						if: {
+							properties: {
+								SERVICE: {
+									const: 'SERVICE_B',
+								},
+							},
 						},
-						'then': {
-							'required': [
-								'SERVICE_B_URL',
-							]
-						}
-					}
-				}
+						then: {
+							required: ['SERVICE_B_URL'],
+						},
+					},
+				},
 			};
 			const validate = ajv.compile(schema);
 
@@ -597,30 +668,30 @@ describe('test configuration', () => {
 
 			it('SERVICE_A url valid', () => {
 				const valid = validate({
-					'SERVICE': 'SERVICE_A',
-					'SERVICE_A_URL': 'http://asd.de',
+					SERVICE: 'SERVICE_A',
+					SERVICE_A_URL: 'http://asd.de',
 				});
 				expect(valid).to.be.true;
 			});
 
 			it('SERVICE_A url missing', () => {
 				const valid = validate({
-					'SERVICE': 'SERVICE_A',
+					SERVICE: 'SERVICE_A',
 				});
 				expect(valid).to.be.false;
 			});
 
 			it('SERVICE_B url valid', () => {
 				const valid = validate({
-					'SERVICE': 'SERVICE_B',
-					'SERVICE_B_URL': 'http://asd.de',
+					SERVICE: 'SERVICE_B',
+					SERVICE_B_URL: 'http://asd.de',
 				});
 				expect(valid).to.be.true;
 			});
 
 			it('SERVICE_B url missing', () => {
 				const valid = validate({
-					'SERVICE': 'SERVICE_B',
+					SERVICE: 'SERVICE_B',
 				});
 				expect(valid).to.be.false;
 			});